+++ src/apple/apple-wizard.ts
@@ -10,9 +10,11 @@ import * as path from 'path';
 import * as xcManager from './xcode-manager';
 import * as codeTools from './code-tools';
 import * as bash from '../utils/bash';
-import { WizardOptions } from '../utils/types';
+import { SentryProjectData, WizardOptions } from '../utils/types';
 import * as Sentry from '@sentry/node';
-import { traceStep } from '../telemetry';
+import { traceStep, withTelemetry } from '../telemetry';
+import * as cocoapod from './cocoapod';
+import * as fastlane from './fastlane';
 
 const xcode = require('xcode');
 /* eslint-enable @typescript-eslint/no-unused-vars */
@@ -22,13 +24,22 @@ import {
   askForSelfHosted,
   askForWizardLogin,
   askToInstallSentryCLI,
-  SentryProjectData,
   printWelcome,
   abort,
   askForItemSelection,
 } from '../utils/clack-utils';
 
-export async function runAppleWizard(
+export async function runAppleWizard(options: WizardOptions): Promise<void> {
+  return withTelemetry(
+    {
+      enabled: options.telemetryEnabled,
+      integration: 'ios',
+    },
+    () => runAppleWizardWithTelementry(options),
+  );
+}
+
+async function runAppleWizardWithTelementry(
   options: WizardOptions,
 ): Promise<void> {
   printWelcome({
@@ -39,8 +50,12 @@ export async function runAppleWizard(
   const hasCli = bash.hasSentryCLI();
   Sentry.setTag('has-cli', hasCli);
   if (!hasCli) {
-    if (!(await traceStep("Ask for SentryCLI", () => askToInstallSentryCLI()))) {
-      clack.log.warn("Without sentry-cli, you won't be able to upload debug symbols to Sentry. You can install it later by following the instructions at https://docs.sentry.io/cli/");
+    if (
+      !(await traceStep('Ask for SentryCLI', () => askToInstallSentryCLI()))
+    ) {
+      clack.log.warn(
+        "Without sentry-cli, you won't be able to upload debug symbols to Sentry. You can install it later by following the instructions at https://docs.sentry.io/cli/",
+      );
       Sentry.setTag('CLI-Installed', false);
     } else {
       await bash.installSentryCLI();
@@ -49,10 +64,12 @@ export async function runAppleWizard(
   }
 
   const projectDir = process.cwd();
-  const xcodeProjFiles = findFilesWithExtension(projectDir, ".xcodeproj");
+  const xcodeProjFiles = findFilesWithExtension(projectDir, '.xcodeproj');
 
   if (!xcodeProjFiles || xcodeProjFiles.length === 0) {
-    clack.log.error('No Xcode project found. Please run this command from the root of your project.');
+    clack.log.error(
+      'No Xcode project found. Please run this command from the root of your project.',
+    );
     await abort();
     return;
   }
@@ -64,34 +81,88 @@ export async function runAppleWizard(
     Sentry.setTag('multiple-projects', false);
   } else {
     Sentry.setTag('multiple-projects', true);
-    xcodeProjFile = (await traceStep("Choose Xcode project", () => askForItemSelection(xcodeProjFiles, "Which project do you want to add Sentry to?"))).value;
+    xcodeProjFile = (
+      await traceStep('Choose Xcode project', () =>
+        askForItemSelection(
+          xcodeProjFiles,
+          'Which project do you want to add Sentry to?',
+        ),
+      )
+    ).value;
   }
 
-  const pbxproj = path.join(projectDir, xcodeProjFile, "project.pbxproj");
+  const pbxproj = path.join(projectDir, xcodeProjFile, 'project.pbxproj');
   if (!fs.existsSync(pbxproj)) {
     clack.log.error(`No pbxproj found at ${pbxproj}`);
     await abort();
     return;
   }
 
-  const { project, apiKey } = await getSentryProjectAndApiKey(options.promoCode, options.url);
+  const { project, apiKey } = await getSentryProjectAndApiKey(
+    options.promoCode,
+    options.url,
+  );
+
+  const hasCocoa = cocoapod.usesCocoaPod(projectDir);
+
+  if (hasCocoa) {
+    const podAdded = await traceStep('Add CocoaPods reference', () =>
+      cocoapod.addCocoaPods(projectDir),
+    );
+    if (!podAdded) {
+      clack.log.warn(
+        "Could not add Sentry pod to your Podfile. You'll have to add it manually.\nPlease follow the instructions at https://docs.sentry.io/platforms/apple/guides/ios/#install",
+      );
+    }
+  }
 
   traceStep('Update Xcode project', () => {
-    xcManager.updateXcodeProject(pbxproj, project, apiKey, true, true);
+    xcManager.updateXcodeProject(pbxproj, project, apiKey, !hasCocoa, true);
   });
 
-  const projSource = path.join(projectDir, xcodeProjFile.replace(".xcodeproj", ""));
-  const codeAdded = traceStep("Add code snippet", () => { return codeTools.addCodeSnippetToProject(projSource, project.keys[0].dsn.public) });
+  Sentry.setTag('package-manager', hasCocoa ? 'cocoapods' : 'SPM');
+  const projSource = path.join(
+    projectDir,
+    xcodeProjFile.replace('.xcodeproj', ''),
+  );
+  const codeAdded = traceStep('Add code snippet', () => {
+    return codeTools.addCodeSnippetToProject(
+      projSource,
+      project.keys[0].dsn.public,
+    );
+  });
   if (!codeAdded) {
-    clack.log.warn('Added the Sentry dependency to your project but could not add the Sentry code snippet. Please add the code snipped manually by following the docs: https://docs.sentry.io/platforms/apple/guides/ios/#configure');
+    clack.log.warn(
+      'Added the Sentry dependency to your project but could not add the Sentry code snippet. Please add the code snipped manually by following the docs: https://docs.sentry.io/platforms/apple/guides/ios/#configure',
+    );
     return;
+  }
+
+  if (fastlane.fastFile(projectDir)) {
+    const addLane = await clack.confirm({
+      message:
+        'Found a Fastfile in your project. Do you want to configure a lane to upload debug symbols to Sentry?',
+    });
+    if (addLane) {
+      await traceStep('Configure fastlane', () =>
+        fastlane.addSentryToFastlane(
+          projectDir,
+          project.organization.slug,
+          project.slug,
+          apiKey.token,
+        ),
+      );
+    }
   }
 
   clack.log.success('Sentry was successfully added to your project!');
 }
 
 //Prompt for Sentry project and API key
-async function getSentryProjectAndApiKey(promoCode?: string, url?: string): Promise<{ project: SentryProjectData, apiKey: { token: string } }> {
+async function getSentryProjectAndApiKey(
+  promoCode?: string,
+  url?: string,
+): Promise<{ project: SentryProjectData; apiKey: { token: string } }> {
   const { url: sentryUrl } = await askForSelfHosted(url);
 
   const { projects, apiKeys } = await askForWizardLogin({
@@ -107,5 +178,5 @@ async function getSentryProjectAndApiKey(promoCode?: string, url?: string): Prom
 //find files with the given extension
 function findFilesWithExtension(dir: string, extension: string): string[] {
   const files = fs.readdirSync(dir);
-  return files.filter(file => file.endsWith(extension));
+  return files.filter((file) => file.endsWith(extension));
 }