+++ src/apple/templates.ts
@@ -1,17 +1,20 @@
 export function getRunScriptTemplate(
-    orgSlug: string,
-    projectSlug: string,
-    apiKey: string,
-    uploadSource = true,
+  orgSlug: string,
+  projectSlug: string,
+  apiKey: string,
+  uploadSource = true,
 ): string {
-    // eslint-disable-next-line no-useless-escape
-    return `# This script is responsable to upload debug symbols and source context for Sentry.\\nif which sentry-cli >/dev/null; then\\nexport SENTRY_ORG=${orgSlug}\\nexport SENTRY_PROJECT=${projectSlug}\\nexport SENTRY_AUTH_TOKEN=${apiKey}\\nERROR=$(sentry-cli debug-files upload ${uploadSource ? "--include-sources " : ""}\"$DWARF_DSYM_FOLDER_PATH\" 2>&1 >/dev/null)\\nif [ ! $? -eq 0 ]; then\\necho \"warning: sentry-cli - $ERROR\"\\nfi\\nelse\\necho \"warning: sentry-cli not installed, download from https://github.com/getsentry/sentry-cli/releases\"\\nfi\\n`;
+  // eslint-disable-next-line no-useless-escape
+  return `# This script is responsable to upload debug symbols and source context for Sentry.\\nif which sentry-cli >/dev/null; then\\nexport SENTRY_ORG=${orgSlug}\\nexport SENTRY_PROJECT=${projectSlug}\\nexport SENTRY_AUTH_TOKEN=${apiKey}\\nERROR=$(sentry-cli debug-files upload ${
+    uploadSource ? '--include-sources ' : ''
+  }"$DWARF_DSYM_FOLDER_PATH" 2>&1 >/dev/null)\\nif [ ! $? -eq 0 ]; then\\necho "warning: sentry-cli - $ERROR"\\nfi\\nelse\\necho "warning: sentry-cli not installed, download from https://github.com/getsentry/sentry-cli/releases"\\nfi\\n`;
 }
 
-export const scriptInputPath = "\"${DWARF_DSYM_FOLDER_PATH}/${DWARF_DSYM_FILE_NAME}/Contents/Resources/DWARF/${TARGET_NAME}\"";
+export const scriptInputPath =
+  '"${DWARF_DSYM_FOLDER_PATH}/${DWARF_DSYM_FILE_NAME}/Contents/Resources/DWARF/${TARGET_NAME}"';
 
 export function getSwiftSnippet(dsn: string): string {
-    return `        SentrySDK.start { options in
+  return `        SentrySDK.start { options in
             options.dsn = "${dsn}"
             options.debug = true // Enabled debug when first installing is always helpful
             options.enableTracing = true 
@@ -25,7 +28,7 @@ export function getSwiftSnippet(dsn: string): string {
 }
 
 export function getObjcSnippet(dsn: string): string {
-    return `    [SentrySDK startWithConfigureOptions:^(SentryOptions * options) {
+  return `    [SentrySDK startWithConfigureOptions:^(SentryOptions * options) {
         options.dsn = @"${dsn}";
         options.debug = YES; // Enabled debug when first installing is always helpful
         options.enableTracing = YES;
@@ -36,4 +39,17 @@ export function getObjcSnippet(dsn: string): string {
     }];
     //Remove the next line after confirming that your Sentry integration is working.
     [SentrySDK captureMessage:@"This app uses Sentry!"];\n`;
+}
+
+export function getFastlaneSnippet(
+  org: string,
+  project: string,
+  token: string,
+): string {
+  return `    sentry_cli(
+      auth_token: '${token}',
+      org_slug: '${org}',
+      project_slug: '${project}',
+      include_sources: true
+    )`;
 }