+++ src/nextjs/nextjs-wizard.ts
@@ -11,13 +11,12 @@ import {
   abort,
   abortIfCancelled,
   addSentryCliRc,
-  askForProjectSelection,
-  askForSelfHosted,
-  askForWizardLogin,
   confirmContinueEvenThoughNoGitRepo,
   ensurePackageIsInstalled,
+  getOrAskForProjectData,
   getPackageDotJson,
   installPackage,
+  isUsingTypeScript,
   printWelcome,
 } from '../utils/clack-utils';
 import { WizardOptions } from '../utils/types';
@@ -44,29 +43,15 @@ export async function runNextjsWizard(options: WizardOptions): Promise<void> {
   const packageJson = await getPackageDotJson();
   await ensurePackageIsInstalled(packageJson, 'next', 'Next.js');
 
-  const { url: sentryUrl, selfHosted } = await askForSelfHosted(options.url);
-
-  const { projects, apiKeys } = await askForWizardLogin({
-    promoCode: options.promoCode,
-    url: sentryUrl,
-    platform: 'javascript-nextjs',
-  });
-
-  const selectedProject = await askForProjectSelection(projects);
+  const { selectedProject, authToken, selfHosted, sentryUrl } =
+    await getOrAskForProjectData(options);
 
   await installPackage({
     packageName: '@sentry/nextjs',
     alreadyInstalled: !!packageJson?.dependencies?.['@sentry/nextjs'],
   });
 
-  let isUsingTypescript = false;
-  try {
-    isUsingTypescript = fs.existsSync(
-      path.join(process.cwd(), 'tsconfig.json'),
-    );
-  } catch {
-    // noop - Default to assuming user is not using typescript
-  }
+  const typeScriptDetected = isUsingTypeScript();
 
   const configVariants = ['server', 'client', 'edge'] as const;
 
@@ -114,7 +99,7 @@ export async function runNextjsWizard(options: WizardOptions): Promise<void> {
 
     if (shouldWriteFile) {
       await fs.promises.writeFile(
-        path.join(process.cwd(), isUsingTypescript ? tsConfig : jsConfig),
+        path.join(process.cwd(), typeScriptDetected ? tsConfig : jsConfig),
         getSentryConfigContents(
           selectedProject.keys[0].dsn.public,
           configVariant,
@@ -122,7 +107,9 @@ export async function runNextjsWizard(options: WizardOptions): Promise<void> {
         { encoding: 'utf8', flag: 'w' },
       );
       clack.log.success(
-        `Created fresh ${chalk.bold(isUsingTypescript ? tsConfig : jsConfig)}.`,
+        `Created fresh ${chalk.bold(
+          typeScriptDetected ? tsConfig : jsConfig,
+        )}.`,
       );
     }
   }
@@ -353,7 +340,7 @@ export async function runNextjsWizard(options: WizardOptions): Promise<void> {
     );
   }
 
-  await addSentryCliRc(apiKeys.token);
+  await addSentryCliRc(authToken);
 
   const mightBeUsingVercel = fs.existsSync(
     path.join(process.cwd(), 'vercel.json'),